refactor(outputgraph): outputGraph と output2Graphs を同じファイルに定義する
