+++ src/index.ts
@@ -1,8 +1,7 @@
 import { DangerDSLType } from '../node_modules/danger/distribution/dsl/DangerDSL';
 import getRenameFiles from './getRenameFiles';
 import getFullGraph from './getFullGraph';
-import output2Graphs from './output2Graphs';
-import outputGraph from './outputGraph';
+import { outputGraph, output2Graphs } from './outputGraph';
 // Provides dev-time type structures for  `danger` - doesn't affect runtime.
 declare let danger: DangerDSLType;
 export declare function message(message: string): void;
