+++ src/outputGraph.ts
@@ -13,7 +13,7 @@ import { DangerDSLType } from 'danger/distribution/dsl/DangerDSL';
 declare let danger: DangerDSLType;
 export declare function markdown(message: string): void;
 
-export default function outputGraph(
+export function outputGraph(
   baseGraph: Graph,
   headGraph: Graph,
   meta: Meta,
@@ -69,6 +69,85 @@ export default function outputGraph(
 
 \`\`\`mermaid
 ${mermaidLines.join('\n')}
+\`\`\`
+
+`);
+}
+
+export async function output2Graphs(
+  baseGraph: Graph,
+  headGraph: Graph,
+  meta: Meta,
+  renamed:
+    | {
+        filename: string;
+        previous_filename: string | undefined;
+      }[]
+    | undefined,
+) {
+  const modified = danger.git.modified_files;
+  const created = danger.git.created_files;
+  const deleted = danger.git.deleted_files;
+  // ファイルの削除またはリネームがある場合は Graph を2つ表示する
+  let tmpBaseGraph = abstraction(
+    extractAbstractionTarget(
+      baseGraph,
+      extractNoAbstractionDirs(
+        [
+          created,
+          deleted,
+          modified,
+          (renamed?.map(diff => diff.previous_filename).filter(Boolean) ??
+            []) as string[],
+        ].flat(),
+      ),
+    ),
+    baseGraph,
+  );
+  tmpBaseGraph = addStatus({ modified, created, deleted }, tmpBaseGraph);
+  // base の書き出し
+  const baseLines: string[] = [];
+  await mermaidify((arg: string) => baseLines.push(arg), tmpBaseGraph, {
+    rootDir: meta.rootDir,
+    LR: true,
+  });
+
+  let tmpHeadGraph = abstraction(
+    extractAbstractionTarget(
+      headGraph,
+      extractNoAbstractionDirs(
+        [
+          created,
+          deleted,
+          modified,
+          (renamed?.map(diff => diff.previous_filename).filter(Boolean) ??
+            []) as string[],
+        ].flat(),
+      ),
+    ),
+    headGraph,
+  );
+  tmpHeadGraph = addStatus({ modified, created, deleted }, tmpHeadGraph);
+  // head の書き出し
+  const headLines: string[] = [];
+  await mermaidify((arg: string) => headLines.push(arg), tmpHeadGraph, {
+    rootDir: meta.rootDir,
+    LR: true,
+  });
+
+  markdown(`
+# TypeScript Graph - Diff
+
+## Base Branch
+
+\`\`\`mermaid
+${baseLines.join('\n')}
+\`\`\`
+
+## Head Branch
+
+\`\`\`mermaid
+${headLines.join('\n')}
 \`\`\`
 
 `);
